\name{ascarr}
          \docType{data}
          \alias{ascarr}
          \title{Banner Versions of Characters}
          \description{
            This data set contains the "banner" arrays of characters to be used by the function \code{\link[cgwtools]{banvax}}.  
          }
          \usage{ascarr}
          \format{The array must be of dimension [6,8,N] for N characters. The \code{dimnames} for the third dimension must be the character matching each such layer.}
         
          \keyword{datasets}