\name{cgwtools-package}
\alias{cgwtools-package}
\alias{cgwtools}
\docType{package}
\title{ A collection of tools that the author finds handy}
\description{
Most of these tools are small functions to do simple tasks or filtered views of the current environment.  In addition the function \code{splatnd} is provided primarily as a piece of example code to show how to write zero-argument operators. It's based on the code in the package  \code{sos} , and avoids the need to use \code{makeActiveBinding} (as in ,e.g., \code{pracma::ans}  ) 
}
\details{
\tabular{ll}{
Package: \tab cgwtools\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2014-11-25\cr
License: \tab GPL-3 \cr
}

}
\author{
Carl Witthoft, with attributions as noted in the individual help pages

Maintainer:Carl Witthoft carl@witthoft.com

}

